fix: resolve circular import by moving VespaDocumentType to separate types module
