+++ services/vespa_loader/pubsub_consumer.py
@@ -19,10 +19,8 @@ from services.vespa_loader.settings import Settings
 from services.common.events.email_events import EmailBackfillEvent, EmailData
 from services.common.events.calendar_events import CalendarUpdateEvent, CalendarBatchEvent, CalendarEventData
 from services.common.events.contact_events import ContactUpdateEvent, ContactBatchEvent, ContactData
+from services.vespa_loader.types import VespaDocumentType
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from typing import Any, List, Optional
-from datetime import datetime
 
 # PubSub types
 try:
@@ -34,48 +32,7 @@ except ImportError:
     ReceivedMessage = Any  # type: ignore
 
 
-@dataclass
-class VespaDocumentType:
-    """Typed document structure for Vespa ingestion"""
-    id: str
-    user_id: str
-    type: str
-    provider: str
-    subject: str
-    body: str
-    from_address: str
-    to_addresses: List[str]
-    thread_id: Optional[str] = None
-    folder: Optional[str] = None
-    created_at: Optional[datetime] = None
-    updated_at: Optional[datetime] = None
-    metadata: Optional[Dict[str, Any]] = None
-    content_chunks: Optional[List[str]] = None
-    quoted_content: Optional[str] = None
-    thread_summary: Optional[Dict[str, Any]] = None
-    search_text: Optional[str] = None
 
-    def to_dict(self) -> Dict[str, Any]:
-        """Convert to dictionary format expected by Vespa"""
-        return {
-            "id": self.id,
-            "user_id": self.user_id,
-            "type": self.type,
-            "provider": self.provider,
-            "subject": self.subject,
-            "body": self.body,
-            "from": self.from_address,
-            "to": self.to_addresses,
-            "thread_id": self.thread_id or "",
-            "folder": self.folder or "",
-            "created_at": self.created_at,
-            "updated_at": self.updated_at,
-            "metadata": self.metadata or {},
-            "content_chunks": self.content_chunks or [],
-            "quoted_content": self.quoted_content or "",
-            "thread_summary": self.thread_summary or {},
-            "search_text": self.search_text or "",
-        }
 
 
 logger = get_logger(__name__)
